+++ src/environments/environment.ts
@@ -8,7 +8,7 @@ export const environment = {
   url: 'http://localhost:4200',
   discordURL: 'https://discord.com/invite/xRT6Fz5',
   stripePublicKey: 'pk_test_5D57doatt7FpFolXlTHy8DvZ00TgL3h1WJ',
-  version: 'v2.0.3a',
+  version: 'v2.1.0b',
   test: {
     guildId: '598565371162656788'
   },